Started implementation of dynamic labels on the UI

+++ src/map/dungeon_factory.hpp
@@ -20,7 +20,7 @@ class DungeonFactory {
 public:
     DungeonFactory();
     
-    std::shared_ptr<libpmg::DungeonMap> GenerateDungeon(DungeonCategory const &category, short floor);
+    std::shared_ptr<libpmg::DungeonMap> GenerateDungeon(DungeonCategory const &category, int floor);
     
 private:
     std::unordered_map<DungeonCategory, std::unique_ptr<libpmg::DungeonBuilder>> category_builders_;
